Replaced float with int8 for Ring

+++ collectives.h
@@ -25,11 +25,11 @@ void InitCollectives(int device);
 // The ring allreduce. The lengths of the data chunks passed to this function
 // must be the same across all MPI processes. The output memory will be
 // allocated and written into `output`.
-void RingAllreduce(float* data, size_t length, float** output);
+void RingAllreduce(int8_t* data, size_t length, int8_t** output);
 
 // The ring allgather. The lengths of the data chunks passed to this function
 // may differ across different devices. The output memory will be allocated and
 // written into `output`.
-void RingAllgather(float* data, size_t length, float** output);
+void RingAllgather(int8_t* data, size_t length, int8_t** output);
 
 #endif /* ifndef BAIDU_ALLREDUCE_COLLECTIVES_H_ */
